test: split watch tests to different files
